refactor: change email import path
