+++ api/services/services.user.js
@@ -1,7 +1,7 @@
 const { user, client } = require('../models/models.index')
 const cryptography = require('../utils/utils.cryptography')
-const emailUtils = require('../utils/email/utils.email')
-const { Email } = require('../utils/email/utils.email.recovery')
+const emailUtils = require('../utils/email/email.index')
+const { Email } = require('../utils/email/email.recovery')
 const userMapper = require('../mappers/mappers.user')
 const ErrorGeneric = require('../utils/errors/erros.generic-error')
 const ErrorNotAuthenticatedUser = require('../utils/errors/errors.user-not-authenticated')
